Desktop/nfac-red.ai-main/requirements/base.txt
# Red.AI Backend - Base Requirements
# Основные зависимости для всех окружений

# FastAPI и связанные
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
databases[postgresql]==0.9.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Аутентификация
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# AI/ML сервисы
openai==1.3.7
anthropic==0.7.7
huggingface-hub==0.19.4
torch==2.7.1
torchvision==0.22.1
transformers==4.36.0
diffusers==0.25.0
accelerate==0.25.0
xformers==0.0.31

# Обработка изображений
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
scikit-image==0.22.0

# HTTP клиенты
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Валидация и сериализация
email-validator==2.1.0
python-dateutil==2.8.2
typing-extensions==4.14.1

# Логирование и мониторинг
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Утилиты
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tenacity==8.2.3

# Безопасность
cryptography==41.0.7
bcrypt==4.1.2

# Форматирование и качество кода
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0 